+++ podcasts/End of Year/StoriesModel.swift
@@ -1,26 +1,39 @@
 import Combine
+import PocketCastsServer
 import SwiftUI
 
 @MainActor
 class StoriesModel: ObservableObject {
     var progress: Double
 
-    @Published var currentStory: Int = 0
+    @Published var currentStoryIndex: Int = 0
 
     @Published var isReady: Bool = false
 
     @Published var failed: Bool = false
 
+    @Published var screenshotTaken: Bool = false
+
+    let activeTier: () -> SubscriptionTier
+
     private let dataSource: StoriesDataSource
     private let publisher: Timer.TimerPublisher
     private let configuration: StoriesConfiguration
 
+    private var cancellables = Set<AnyCancellable>()
+
     private var cancellable: Cancellable?
     private var interval: TimeInterval {
-        dataSource.story(for: currentStory).duration
+        currentStory?.duration ?? 0
     }
 
     private var currentStoryIdentifier: String = ""
+
+    private var currentStoryIsPlus = false
+
+    private var manuallyChanged = false
+
+    private var currentStory: Story?
 
     var numberOfStories: Int {
         dataSource.numberOfStories
@@ -30,11 +43,12 @@ class StoriesModel: ObservableObject {
         configuration.storiesToPreload
     }
 
-    init(dataSource: StoriesDataSource, configuration: StoriesConfiguration) {
+    init(dataSource: StoriesDataSource, configuration: StoriesConfiguration, activeTier: @autoclosure @escaping () -> SubscriptionTier = SubscriptionHelper.activeTier) {
         self.dataSource = dataSource
         self.configuration = configuration
         self.progress = 0
         self.publisher = Timer.publish(every: 0.01, on: .main, in: .default)
+        self.activeTier = activeTier
 
         Task.init {
             await isReady = dataSource.isReady()
@@ -42,6 +56,15 @@ class StoriesModel: ObservableObject {
         }
 
         subscribeToNotifications()
+    }
+
+    func refresh() {
+        isReady = false
+
+        Task.init {
+            await isReady = dataSource.refresh()
+            failed = !isReady
+        }
     }
 
     func start() {
@@ -57,29 +80,46 @@ class StoriesModel: ObservableObject {
             if currentStory >= self.numberOfStories || newProgress < 0 {
                 if self.configuration.startOverFromBeginningAfterFinished {
                     newProgress = 0
-                    self.currentStory = 0
+                    self.currentStoryIndex = 0
                 }
                 else {
                     self.pause()
                 }
-            } else if currentStory != self.currentStory {
-                self.currentStory = currentStory
+            } else if currentStory != self.currentStoryIndex {
+                if self.shouldSkipPlusStories() {
+                    newProgress = Double(currentStory + self.numberOfPlusStoriesAfterTheCurrentOne())
+                    self.currentStoryIndex = currentStory + self.numberOfPlusStoriesAfterTheCurrentOne()
+                } else {
+                    self.currentStoryIndex = currentStory
+                }
+                self.manuallyChanged = false
             }
 
             self.progress = newProgress
             StoriesProgressModel.shared.progress = newProgress
         })
+
+        currentStory?.onResume()
     }
 
     func story(index: Int) -> AnyView {
         let story = dataSource.story(for: index)
-        story.onAppear()
+
+        // Only trigger onAppear if the story is not plus or user is paid
+        // Otherwise, the paywall appears in front of the story
+        if !story.plusOnly || isPaidUser() {
+            story.onAppear()
+        }
+
+        currentStory = story
+
         currentStoryIdentifier = story.identifier
+        currentStoryIsPlus = story.plusOnly
         return AnyView(story)
     }
 
-    func storyIsShareable(index: Int) -> Bool {
-        dataSource.shareableStory(for: index) != nil ? true : false
+    func showShareButton(index: Int) -> Bool {
+        dataSource.shareableStory(for: index)?.hideShareButton() == false
     }
 
     func preload(index: Int) -> AnyView {
@@ -91,7 +131,7 @@ class StoriesModel: ObservableObject {
     }
 
     func sharingAssets() -> [Any]? {
-        guard let story = dataSource.shareableStory(for: currentStory) else {
+        guard let story = currentStory as? (any ShareableStory) else {
             return nil
         }
 
@@ -112,28 +152,67 @@ class StoriesModel: ObservableObject {
     }
 
     func next() {
-        guard numberOfStories > 0 else {
+        guard isReady, numberOfStories > 0 else {
             return
         }
 
-        progress = min(Double(numberOfStories), Double(Int(progress) + 1))
+        let nextNonPlus = currentStoryIsPlus ? Int(progress.rounded(.down)) + numberOfPlusStoriesAfterTheCurrentOne() + 1 : 0
+
+        manuallyChanged = true
+
+        progress = min(Double(numberOfStories), Double(max(nextNonPlus, Int(progress) + 1)))
     }
 
     func previous() {
-        guard numberOfStories > 0 else {
+        guard isReady, numberOfStories > 0 else {
             return
         }
 
-        progress = max(0, Double(Int(progress) - 1))
+        let previousNonPlus = currentStoryIndex - numberOfPlusStoriesBeforeTheCurrentOne()
+
+        manuallyChanged = true
+
+        progress = max(0, Double(min(previousNonPlus, Int(progress) - 1)))
+    }
+
+    func numberOfPlusStoriesBeforeTheCurrentOne() -> Int {
+        guard !isPaidUser() else {
+            return 0
+        }
+
+        var currentStory = currentStoryIndex
+        var numberOfStoriesToSkip = 0
+        while currentStory > 0 && dataSource.story(for: currentStory - 1).plusOnly {
+            numberOfStoriesToSkip += 1
+            currentStory -= 1
+        }
+
+        return numberOfStoriesToSkip
+    }
+
+    func numberOfPlusStoriesAfterTheCurrentOne() -> Int {
+        guard !isPaidUser() else {
+            return 0
+        }
+
+        var currentStory = currentStoryIndex
+        var numberOfStoriesToSkip = 0
+        while currentStory + 1 < numberOfStories && dataSource.story(for: currentStory + 1).plusOnly {
+            numberOfStoriesToSkip += 1
+            currentStory += 1
+        }
+
+        return numberOfStoriesToSkip
     }
 
     func pause() {
         cancellable?.cancel()
+        currentStory?.onPause()
     }
 
     func replay() {
         progress = 0
-        currentStory = 0
+        currentStoryIndex = 0
         pause()
         start()
     }
@@ -151,6 +230,10 @@ class StoriesModel: ObservableObject {
         pause()
         NavigationManager.sharedManager.dismissPresentedViewController()
     }
+
+    func shouldShowUpsell() -> Bool {
+        currentStoryIsPlus && activeTier() == .none
+    }
 }
 
 private extension StoriesModel {
@@ -163,5 +246,37 @@ private extension StoriesModel {
                 }
             }
         }
+
+        UIApplication.userDidTakeScreenshotNotification.publisher()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                self?.pause()
+                self?.screenshotTaken = true
+            }
+            .store(in: &cancellables)
+
+        ServerNotifications.iapPurchaseCompleted.publisher()
+        .receive(on: DispatchQueue.main)
+        .sink { [weak self] _ in
+            self?.refresh()
+        }
+        .store(in: &cancellables)
+    }
+
+    func isPaidUser() -> Bool {
+        activeTier() != .none
+    }
+
+    func nextStoryIsPlus() -> Bool {
+        if currentStoryIndex + 1 < numberOfStories {
+            return dataSource.story(for: currentStoryIndex + 1).plusOnly
+        }
+
+        return false
+    }
+
+    /// Whether some Plus stories should be skipped or not
+    func shouldSkipPlusStories() -> Bool {
+        !isPaidUser() && !manuallyChanged && currentStoryIsPlus && nextStoryIsPlus()
     }
 }