+++ podcasts/ProfileViewController.swift
@@ -39,7 +39,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     private let settingsCellId = "SettingsCell"
     private let endOfYearPromptCell = "EndOfYearPromptCell"
 
-    private enum TableRow { case allStats, downloaded, starred, listeningHistory, uploadedFiles, endOfYearPrompt }
+    private enum TableRow { case allStats, downloaded, starred, listeningHistory, help, uploadedFiles, endOfYearPrompt }
 
     @IBOutlet var profileTable: UITableView! {
         didSet {
@@ -124,7 +124,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
             NotificationCenter.postOnMainThread(notification: Constants.Notifications.profileSeen)
         }
 
-        showGeneralSettingsIfNeeded()
+        whatsNewDismissed()
     }
 
     override func viewDidDisappear(_ animated: Bool) {
@@ -270,6 +270,9 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         case .listeningHistory:
             cell.settingsImage.image = UIImage(named: "profile-history")
             cell.settingsLabel.text = L10n.listeningHistory
+        case .help:
+            cell.settingsImage.image = UIImage(named: "profile-help")
+            cell.settingsLabel.text = L10n.settingsHelp
         case .endOfYearPrompt:
             return EndOfYearPromptCell()
         }
@@ -305,6 +308,9 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         case .listeningHistory:
             let historyController = ListeningHistoryViewController()
             navigationController?.pushViewController(historyController, animated: true)
+        case .help:
+            let navController = SJUIUtils.navController(for: OnlineSupportController())
+            present(navController, animated: true, completion: nil)
         case .endOfYearPrompt:
             Analytics.track(.endOfYearProfileCardTapped)
             EndOfYear().showStories(in: self, from: .profile)
@@ -325,7 +331,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
     private func tableData() -> [[ProfileViewController.TableRow]] {
         var data: [[ProfileViewController.TableRow]]
-        data = [[.allStats, .downloaded, .uploadedFiles, .starred, .listeningHistory]]
+        data = [[.allStats, .downloaded, .uploadedFiles, .starred, .listeningHistory, .help]]
 
         if EndOfYear.isEligible {
             data[0].insert(.endOfYearPrompt, at: 0)
@@ -343,19 +349,23 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
     @objc private func whatsNewDismissed() {
         showGeneralSettingsIfNeeded()
+        showHeadphoneControlsFromWhatsNew()
     }
 
     private func showGeneralSettingsIfNeeded() {
-        if AnnouncementFlow.shared.isShowingAutoplayOption {
+        if AnnouncementFlow.current == .autoPlay {
             let generalSettingsViewController = GeneralSettingsViewController()
             navigationController?.pushViewController(generalSettingsViewController, animated: true)
         }
+    }
 
-        if AnnouncementFlow.shared.isShowingBookmarksOption {
-            let controller = HeadphoneSettingsViewController()
-            navigationController?.pushViewController(controller, animated: true)
-            AnnouncementFlow.shared.isShowingBookmarksOption = false
-        }
+    // Pushes to the headphone controls if shown from the what's new
+    private func showHeadphoneControlsFromWhatsNew() {
+        guard AnnouncementFlow.current == .bookmarksProfile else { return }
+
+        let controller = HeadphoneSettingsViewController()
+        navigationController?.pushViewController(controller, animated: true)
+        AnnouncementFlow.current = .none
     }
 }