+++ podcasts/Whats New/Announcements.swift
@@ -18,50 +18,17 @@ struct Announcements {
             message: L10n.announcementAutoplayDescription,
             buttonTitle: L10n.enableItNow,
             action: {
-                AnnouncementFlow.shared.isShowingAutoplayOption = true
+                AnnouncementFlow.current = .autoPlay
 
                 NavigationManager.sharedManager.navigateTo(NavigationManager.settingsProfileKey, data: nil)
             },
             isEnabled: true
         ),
 
-        // Bookmarks Early Access: Beta
-        // Show only in TestFlight, for Plus and Patron
-        .init(
-            version: "99.99",
-            header: AnyView(BookmarksWhatsNewHeader()),
-            title: L10n.announcementBookmarksTitleBeta,
-            message: L10n.announcementBookmarksDescription,
-            buttonTitle: bookmarksViewModel.enableButtonTitle,
-            action: {
-                bookmarksViewModel.enableAction()
-            },
-            displayTier: bookmarksViewModel.displayTier,
-            isEnabled: bookmarksViewModel.isEarlyAccessBetaAnnouncementEnabled
-        ),
-
-        // Bookmarks Early Access: Release
-        // Show when not in beta, for Patron only
-        .init(
-            version: "99.99",
-            header: AnyView(BookmarksWhatsNewHeader().onAppear {
-                // Record when someone sees the full announcement while in early access so we don't show it again to them when we move to full release.
-                bookmarksViewModel.markAsSeen()
-            }),
-            title: L10n.announcementBookmarksTitle,
-            message: L10n.announcementBookmarksDescription,
-            buttonTitle: bookmarksViewModel.enableButtonTitle,
-            action: {
-                bookmarksViewModel.enableAction()
-            },
-            displayTier: bookmarksViewModel.displayTier,
-            isEnabled: bookmarksViewModel.isEarlyAccessAnnouncementEnabled
-        ),
-
         // Bookmarks: Full Release
         // Show for everyone, except those who saw the `Early Access: Release` announcement
         .init(
-            version: "99.99",
+            version: "7.53",
             header: AnyView(BookmarksWhatsNewHeader()),
             title: L10n.announcementBookmarksTitle,
             message: L10n.announcementBookmarksDescription,
@@ -75,9 +42,20 @@ struct Announcements {
     ]
 }
 
-class AnnouncementFlow {
-    static let shared = AnnouncementFlow()
+// MARK: - AnnouncementFlow
 
-    var isShowingAutoplayOption = false
-    var isShowingBookmarksOption = false
+enum AnnouncementFlow {
+    static var current: Self = .none
+
+    /// No active flow
+    case none
+
+    /// Show the autoplay settings
+    case autoPlay
+
+    /// Show the player and highlight the Add Bookmark item
+    case bookmarksPlayer
+
+    /// Show the headphone controls action for Bookmarks
+    case bookmarksProfile
 }